refactor: move accessible button
