+++ packages/components/buttons/accessible-button/src/accessible-button.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../../../src/test-utils';
 import AccessibleButton from './accessible-button';
 
 const createTestProps = custom => ({
