+++ src/components/buttons/primary-button/primary-button.js
@@ -7,7 +7,7 @@ import { css } from '@emotion/core';
 import Inline from '@commercetools-uikit/spacings-inline';
 import { customProperties as vars } from '@commercetools-uikit/design-system';
 import { filterInvalidAttributes } from '@commercetools-uikit/utils';
-import AccessibleButton from '../accessible-button';
+import AccessibleButton from '@commercetools-uikit/accessible-button';
 import { getButtonStyles } from './primary-button.styles';
 
 const propsToOmit = ['type'];
